Show coordinates on the VSD when designing.
As you move the mouse over the image you should see the x y coordinates
in a small box in the upper right. If you click and drag it will
show you both the x y coordinate and the width and height.

It will also copy them onto your clipboard in a format that be perfect
for pasting into 4 consecutive columns of your spreadsheet.